Structure refactoring 1
