+++ app/__init__.py
@@ -10,7 +10,7 @@ def create_app(config_name='default'):
     os.makedirs(app.config['UPLOAD_FOLDER'], exist_ok=True)
     
     # Import and register blueprints for features
-    from app.features.pdf_processor.views import pdf_processor_bp
+    from app.features.bullet_document_report.views import pdf_processor_bp
     from app.features.excel_processor.views import excel_processor_bp
     
     app.register_blueprint(pdf_processor_bp)
