+++ app/features/bullet_document_report/__init__.py
@@ -1,0 +1,2 @@
+# Bullet document report module.
+# THIS MODULE IS FOR BULLET DOCUMENT REPORTS.
