+++ app/features/bullet_document_report/template/js/pdf_processor.js
