--- app/features/pdf_processor/__init__.py
@@ -1,1 +1,0 @@
-# PDF Processor feature module
