+++ app/infra/__init__.py
@@ -1,0 +1,1 @@
+# Infra module
