+++ app/infra/pdf/__init__.py
@@ -1,0 +1,1 @@
+# Pdf service module
